requirements-test.txt
moto==2.0.8
Pillow==8.2.0
pytest
pytest-mock
pytest-html
boto3